+++ app/src/main/res/layout/fragment_mission.xml
@@ -5,8 +5,6 @@
 
     <data>
 
-        <import type="androidx.core.content.ContextCompat" />
-
         <variable
             name="viewmodel"
             type="org.sopt.zooczoocbbangbbang.presentation.main.record.MissionViewModel" />
@@ -15,25 +13,27 @@
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/cl_mission_bg"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:minHeight="627dp"
         tools:context=".presentation.main.record.MissionFragment">
 
         <androidx.viewpager2.widget.ViewPager2
             android:id="@+id/fcv_mission_view"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
+            android:layout_height="0dp"
             android:name="org.sopt.zooczoocbbangbbang.presentation.main.record.MissionViewPagerFragment"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toTopOf="@+id/btn_mission_bottom" />
 
         <androidx.appcompat.widget.AppCompatButton
             android:id="@+id/btn_mission_bottom"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginHorizontal="30dp"
-            android:layout_marginTop="69dp"
             android:clickable="@{viewmodel.buttonValidation}"
             android:background="@{viewmodel.buttonValidation ? @drawable/shape_green_radius_47 : @drawable/shape_grey_radius_47}"
             android:elevation="10dp"
@@ -43,7 +43,8 @@
             android:textAppearance="@style/Typography.S20"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/fcv_mission_view" />
+            app:layout_constraintBottom_toBottomOf="parent" />
+
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
